Implement analytics, persistence APIs, SSR wrappers, and image optimization
