+++ lib/server/jsonStore.ts
@@ -1,0 +1,26 @@
+import { promises as fs } from 'fs';
+import path from 'path';
+
+const DATA_DIR = path.join(process.cwd(), 'data');
+
+async function ensureDataDir(): Promise<void> {
+  await fs.mkdir(DATA_DIR, { recursive: true });
+}
+
+export async function readJsonFile<T>(fileName: string, fallback: T): Promise<T> {
+  await ensureDataDir();
+  const filePath = path.join(DATA_DIR, fileName);
+
+  try {
+    const raw = await fs.readFile(filePath, 'utf8');
+    return JSON.parse(raw) as T;
+  } catch {
+    return fallback;
+  }
+}
+
+export async function writeJsonFile<T>(fileName: string, value: T): Promise<void> {
+  await ensureDataDir();
+  const filePath = path.join(DATA_DIR, fileName);
+  await fs.writeFile(filePath, JSON.stringify(value, null, 2), 'utf8');
+}
